boost tasks-api

+++ staff/manuel-barzi/tasks/tasks-api/logic/register-user/index.js
@@ -1,25 +1,31 @@
 const validate = require('../../utils/validate')
-const users = require('../../data/users.json')
+const users = require('../../data/users')
 const fs = require('fs')
 const path = require('path')
+const uuid = require('uuid/v4')
 
-module.exports = function(name, surname, email, username, password) {
+module.exports = function (name, surname, email, username, password) {
     validate.string(name)
     validate.string.notVoid('name', name)
     validate.string(surname)
     validate.string.notVoid('surname', surname)
     validate.string(email)
     validate.string.notVoid('e-mail', email)
+    validate.email(email)
     validate.string(username)
     validate.string.notVoid('username', username)
     validate.string(password)
     validate.string.notVoid('password', password)
 
     return new Promise((resolve, reject) => {
-        users.push({ name, surname, email, username, password })
+        const user = users.find(user => user.username === username)
 
-        // logic rules!!!!
+        if (user) return reject(Error(`user with username ${username} already exists`))
 
-        fs.writeFile(path.join(__dirname, '../../data/users.json'), JSON.stringify(users), error => error? reject(error) : resolve())
+        const id = uuid()
+
+        users.push({ id, name, surname, email, username, password })
+
+        fs.writeFile(path.join(__dirname, '../../data/users.json'), JSON.stringify(users), error => error ? reject(error) : resolve())
     })
 }
